+++ src/index.ts
@@ -3,3 +3,4 @@ export { Release } from './Release'
 
 export { parseChangelog } from './parseChangelog'
 export { printChangelog } from './printChangelog'
+export { printRelease } from './printRelease'